customRates organized
