+++ src/react/components/Cards/RatesPairList.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import {CardText,CardList} from '../StoryBook'
-import RatesCard from '../RatesCard'
+import {CardText,CardList} from './index'
+import RatesCard from './RatesCard'
 
 
 /*Part of Rates component */
